// src/data/node-types/motor-shows/types/ApiMotorShowHasImageRelationship.ts
import type {ApiRelationshipType} from "../../../types/ApiRelationshipType"
import type {ApiNodeType} from "../../../types/ApiNodeType"
import type {ApiImageNode} from "../../images/types/ApiImageNode"

export type ApiMotorShowHasImageRelationship = {
    data: {
        relationship_id: number
        relationship_name: ApiRelationshipType.MOTOR_SHOW_HAS_IMAGE
        partner_node: {
            node_type: ApiNodeType.IMAGE
            data: ApiImageNode['attributes']
        }
        created_at: string
        updated_at: string
    }
}